Add more metadata to pom.xml, according to Sonatype requirements
Maven central is strict about requirements for a project before letting
it be published:

http://central.sonatype.org/pages/requirements.html

This adds the required metadata (name/description/url/license/repo/owners).
It also sets up a Maven profile "open-source" which overrides distribution
settings. When we "mvn deploy -Popen-source,release", the project we are
building will not be deployed to the NAV internal Nexus, but rather
directly to Maven Central. It will also GPG sign all the artifacts,
which is a security measure required by Maven Central.