Extract Modig components into time capsule module
The Modig library is largely unmaintained. It is not published on
Github, and depends on too many external libraries. It is largely viewed
as a blocker to being able to compile software outside of NAV's test
environment.

This commit extracts the parts of Modig used by the common libraries
into a separate module called `modig-backport`.

* abac: test dependencies on subject handler
* api-app: the Jetty subject handler
* cxf: SAML token handling