[BREAKING] replacing modig-security and oidc-security with new auth module
Implementing new authentication module providing a SubjectHandler that replaces previous implementations in modig-security and oidc-security.
Unlike previous implementations, the new SubjectHandler does not rely on container login or java EE apis. This change attempts to provide
a clean migration path for both older jboss/was-applications and new nais-applications. However if required, integrating modig-security with the
new SubjectHandler is fairly easy.