Fixing resolving of classpath resources
When we embed jetty, some classes might be loaded by the default WebAppClassLoader and some by the system class loader. These classes will be incompatible with each other. Also, Jetty does not consult the classloader of the webapp when resolving resources such as the swagger-ui. We mitigate both these problems by installing an empty classloader that will always defer to the system classloader.

PUS-161